main + add libft.h

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: mcherel- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/30 16:45:15 by mcherel-          #+#    #+#             */
-/*   Updated: 2021/12/02 14:47:16 by mcherel-         ###   ########.fr       */
+/*   Updated: 2021/12/02 17:10:06 by mcherel-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void				*ft_memset(void * pointer, int value, size_t count);
 void				*ft_memmove(void *dest, const void *src, size_t n);
 void				*ft_memchr(const void *b, int c, size_t n);
 int					ft_memcmp(const void *b1, const void *b2, size_t n);
-int					ft_strlen(const char *s);
+size_t					ft_strlen(const char *s);
 size_t				ft_strlcpy(char *dest, const char *src, size_t dstsize);
 size_t				ft_strlcat(char *dest, const char *src, size_t dstsize);
 char				*ft_strchr(const char *s, int c);
